simplify root.go
